+++ CHANGELOG.rst
@@ -11,6 +11,8 @@ The format is based on `Keep a Changelog <https://keepachangelog.com>`_.
 
 Added
 -----
+- `IlluminaBarcodeParser` to parse barcodes from Illumina data.
+
 - `CodonVariant.numCodonMutsByType` method to get numerical values for codon mutations per variant.
 
 Fixed